Begin rewrite of constitution-js
Currently includes all Constitution contract functionality, but is
untested save for the generic transaction generation and signing logic.
Pool contract logic is entirely missing, as are data validation
functions.

General philosophy here is to wrap around all of the contracts'
functions in a consistent way.
Generating a transaction (offline, synchronous) is decoupled from
checking if a transaction will succeed (online, asynchronous).
While all the relevant logic is provided, it is up to the UI to call
upon this logic when doing input and transaction validation, prior to
transaction generation.

The transaction check functions, in cases where a transaction is
concluded to not be successful, return a reason alongside the false
result. The UI can use this to inform the user about why their desired
transaction is rejected.

Some data retrieval functions are provided, but they do not wrap around
all of the getter provided in the Ships contract. Instead, when using
this library it is encouraged to use the broader getShip() (or
getPoll()) function, rather than more granular reads, to limit the
amount of network traffic (and thus UX lag) that has to be done.