Improve testing and add basic CI setup.
Brings in the constitution as a dependency, and also adds the requisite
dev dependencies for building and deploying the contracts on a local
Ganache node.  Adds several testing hooks for then running the tests via
a simple 'npm test', without the need to manually deal with Ganache.

Also adds a basic Travis config.